requirements.txt
flask-cors==5.0.1
groq==0.25.0
python-dotenv==1.1.0
google-genai==1.3.0
